Fix examples, some refactoring around gen_server, Fix notebook
