+++ examples/e02_registered_process.py
@@ -19,8 +19,6 @@ import logging
 from term import Atom
 from pyrlang.node import Node
 from pyrlang.process import Process
-# from pyrlang import GeventEngine as Engine
-from pyrlang import AsyncioEngine as Engine
 from colors import color
 
 LOG = logging.getLogger(color("EXAMPLE2", fg='lime'))
