+++ pyrlang/gen/__init__.py
@@ -14,4 +14,3 @@
 """
 Implementation of Erlangs OTP Gen behaviours (gen_server for starters)
 """
-from .base import *
