+++ pyrlang/net_kernel.py
@@ -14,10 +14,11 @@
 
 import logging
 
-from pyrlang.gen_server import GenServer
+from pyrlang.gen.server import GenServer
+from pyrlang.gen.decorators import call
 from term.atom import Atom
 
-LOG = logging.getLogger("pyrlang")
+LOG = logging.getLogger(__name__)
 
 
 class NetKernel(GenServer):
@@ -28,12 +29,13 @@ class NetKernel(GenServer):
     def __init__(self) -> None:
         """ :param node: pyrlang.node.Node
         """
-        GenServer.__init__(self,
-                           accepted_calls=['is_auth'])
+        super().__init__()
         self.node_db.get().register_name(self, Atom('net_kernel'))
 
-    @staticmethod
-    def is_auth():
+    @call(lambda msg: type(msg) == tuple and
+                      len(msg) == 2 and
+                      msg[0] == Atom('is_auth'))
+    def is_auth(self, msg):
         return Atom('yes')
